fix : swagger 파일 예시 수정 #5
